LPS-71635 Trasformed DocumentConversionImpl in a gateway to different ConverterWrapper (the default one is OpenOffice) to be able to manage formats not supported by OpenOffice or use alternatives
