+++ modules/apps/collaboration/document-library/document-library-document-conversion/src/main/java/com/liferay/document/library/document/conversion/wrapper/BaseConverterWrapper.java
@@ -1,0 +1,26 @@
+package com.liferay.document.library.document.conversion.wrapper;
+
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.SystemProperties;
+
+public abstract class BaseConverterWrapper implements ConverterWrapper {
+
+	public static String getFilePath(String id, String targetExtension) {
+		StringBundler sb = new StringBundler(5);
+
+		sb.append(SystemProperties.get(SystemProperties.TMP_DIR));
+		sb.append("/liferay/document_conversion/");
+		sb.append(id);
+		sb.append(StringPool.PERIOD);
+		sb.append(targetExtension);
+
+		return sb.toString();
+	}
+
+	@Override
+	public boolean isEnabled() {
+		return false;
+	}
+
+}
